feat: use spytial-core as the Alloy XML parser (single source of truth)
`packages/alloy-instance` now re-exports `spytial-core/alloy-instance`
and drops its duplicated copy of the parser. The only Cope and Drag /
Forge-specific helpers (isOutOfInstances, NO_MORE_INSTANCES_SIG_LABEL,
generateGraphId) move to a small local module; everything else (the
AlloyDatum/instance/type model + parseAlloyXML) comes from spytial-core.

- spytial-core 2.6.0 -> 2.6.4 (adds the importable `spytial-core/alloy-instance`
  module + looplength loopBack + visualizer cnd/theme).
- TypeScript 4.9.5 -> 5.x and moduleResolution: "bundler" so the package
  `exports` subpath resolves under tsc/IDE (webpack already honored it via
  transpileOnly + TsconfigPathsPlugin).

Verified: build:forge succeeds; alloy-instance jest tests pass (9/9) against
the re-export; this change adds no new tsc errors (the ~140 remaining are
pre-existing latent type debt — @graph-ts/*, re-reselect, immer, implicit
anys — surfaced only because the build is transpileOnly and CI never typechecks).
