+++ package.json
@@ -60,7 +60,7 @@
     "react-monaco-editor": "^0.47.0",
     "react-redux": "^7.2.6",
     "redux": "^4.1.2",
-    "spytial-core": "2.6.0",
+    "spytial-core": "2.6.4",
     "transformation-matrix": "^2.10.0",
     "ts-is-present": "^1.2.2",
     "uuid": "^9.0.0"
@@ -113,7 +113,7 @@
     "ts-jest": "^29.0.5",
     "ts-loader": "^9.2.6",
     "tsconfig-paths-webpack-plugin": "^3.5.2",
-    "typescript": "^4.9.5",
+    "typescript": "^5.4.0",
     "webpack": "^5.65.0",
     "webpack-cli": "^6.0.1",
     "webpack-dev-server": "^4.6.0"
